% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{fitted.glmmFELMod}
\alias{fitted.glmmFELMod}
\title{Extract fitted values}
\usage{
\method{fitted}{glmmFELMod}(object, type = c("response", "link"), ...)
}
\arguments{
\item{object}{A \code{glmmFELMod} object.}

\item{type}{Either \code{"link"} (linear predictor) or \code{"response"}.}

\item{...}{Unused.}
}
\value{
A numeric vector of fitted values. If \code{type = "link"}, the fitted
linear predictor values are returned. If \code{type = "response"}, the fitted
mean response values on the response scale are returned. The length equals
the number of observations in the fitted object.
}
\description{
Extract fitted values
}
