% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_helpers.R
\name{glmmfe_pl_objective}
\alias{glmmfe_pl_objective}
\title{vp_cp-style PL objective (includes constants)}
\usage{
glmmfe_pl_objective(
  z_work,
  w_num,
  X,
  Z,
  beta,
  eta,
  tau2,
  inner,
  approx,
  vc_eps = 1e-12
)
}
\description{
vp_cp-style PL objective (includes constants)
}
\keyword{internal}
