% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{predict.glmmFELMod}
\alias{predict.glmmFELMod}
\title{Predict from a fitted glmmFEL model}
\usage{
\method{predict}{glmmFELMod}(object, newdata = NULL, type = c("response", "link"), ...)
}
\arguments{
\item{object}{A \code{glmmFELMod} object.}

\item{newdata}{Not supported in this branch (matrix interface only).}

\item{type}{Either \code{"link"} or \code{"response"}.}

\item{...}{Unused.}
}
\value{
A numeric vector of predictions. If \code{type = "link"}, predictions are
returned on the linear predictor scale. If \code{type = "response"}, predictions
are returned on the response scale. The length equals the number of
observations used to fit the model. Supplying \code{newdata} triggers an error
in this matrix-only branch.
}
\description{
Predict from a fitted glmmFEL model
}
