% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{print.glmmFELMod}
\alias{print.glmmFELMod}
\title{Print a glmmFEL model object}
\usage{
\method{print}{glmmFELMod}(x, ...)
}
\arguments{
\item{x}{A \code{glmmFELMod} object.}

\item{...}{Unused.}
}
\value{
Returns \code{x} invisibly (a \code{glmmFELMod} object), called for its side
effect of printing model information to the console.
}
\description{
Print a glmmFEL model object
}
