% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{print.summary.glmmFELMod}
\alias{print.summary.glmmFELMod}
\title{Print a summary.glmmFELMod object}
\usage{
\method{print}{summary.glmmFELMod}(x, ...)
}
\arguments{
\item{x}{A \code{summary.glmmFELMod} object.}

\item{...}{Unused.}
}
\value{
Returns \code{x} invisibly (a \code{summary.glmmFELMod} object), called for its
side effect of printing the summary to the console.
}
\description{
Print a summary.glmmFELMod object
}
