% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{vcov.glmmFELMod}
\alias{vcov.glmmFELMod}
\title{Extract the covariance matrix of the fixed effects}
\usage{
\method{vcov}{glmmFELMod}(object, ...)
}
\arguments{
\item{object}{A \code{glmmFELMod} object.}

\item{...}{Unused.}
}
\value{
A variance-covariance matrix for the estimated fixed-effect
regression coefficients. Row and column names correspond to coefficient
names when available.
}
\description{
Extract the covariance matrix of the fixed effects
}
