% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmFEL-methods.R
\name{logLik.glmmFELMod}
\alias{logLik.glmmFELMod}
\title{Extract log-likelihood (approximate)}
\usage{
\method{logLik}{glmmFELMod}(object, ...)
}
\arguments{
\item{object}{A \code{glmmFELMod} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{"logLik"} giving the approximate log-likelihood
stored in \code{object$logLik}, with attributes \code{"df"} (effective number of
parameters, taken as \code{length(beta) + 1} for the variance component) and
\code{"nobs"} (number of observations).
}
\description{
Extract log-likelihood (approximate)
}
